[IMP] mgmtsystem_nonconformity: Remove mgmtsystem.nonconformity.abstract
In v18, mgmtsystem.nonconformity.abstract has been removed so that mail.thread can be used directly without defining the fields.:
- https://github.com/OCA/management-system/commit/26785dbd812c3b4ccd7a3f5b460df5f0eb80f274
- https://github.com/OCA/management-system/commit/51a368e33c6f9bea2d522e849ae60e7c348fa1ca
- https://github.com/OCA/management-system/commit/fe25644243f24ef2e3dcc602e34953c717683cb8
- https://github.com/OCA/management-system/commit/51abc028729daad2d0d17af6399822a61115d21d
- https://github.com/OCA/management-system/commit/bd288b77a89dceee93abbb3e15711f00443e0692

TT62128
