+++ mgmtsystem_nonconformity/static/src/components/chatter_topbar/chatter_topbar.esm.js
@@ -1,11 +1,17 @@
-/** @odoo-module **/
-
+/* @odoo-module */
 import {Chatter} from "@mail/core/web/chatter";
+import {ThreadService} from "@mail/core/common/thread_service";
 import {patch} from "@web/core/utils/patch";
 
+patch(ThreadService.prototype, {
+    async fetchData(thread, ...args) {
+        const result = await super.fetchData(thread, ...args);
+        thread.non_conformity_count = result.non_conformity_count;
+        return result;
+    },
+});
 patch(Chatter.prototype, {
     async onClickShowNonConformities() {
-        console.log(this);
         if (this.isTemporary) {
             const saved = await this.doSaveRecord();
             if (!saved) {
