+++ mgmtsystem_nonconformity/static/src/components/chatter_topbar/chatter_topbar.xml
@@ -11,7 +11,7 @@
                 type="button"
                 t-att-disabled="!state.thread.isTemporary and !state.thread.hasWriteAccess"
                 t-on-click="onClickShowNonConformities"
-                t-if="props.webRecord.data.non_conformity_count !== undefined"
+                t-if="state.thread.non_conformity_count"
             >
                 <i
                     class="fa fa-exclamation-triangle fa-lg me-1"
@@ -20,7 +20,7 @@
                 />
                <span
                     class="o_ChatterTopbar_buttonCount o_ChatterTopbar_buttonNonConformitiesCount"
-                    t-esc="props.webRecord.data.non_conformity_count"
+                    t-esc="state.thread.non_conformity_count"
                 />
             </button>
         </xpath>
